requirements.txt
langchain_community
langchain-cli
streamlit
streamlit_chat
chromadb
pypdf
fastembed
sentence-transformers